perf: reduce memory use in storing eols

+++ tools/Bench/TokenizationBench.cs
@@ -44,7 +44,7 @@ public class TokenizationBench
 
     private static readonly uint[] _fieldBuffer = new uint[24 * 65535];
     private static readonly byte[] _quoteBuffer = new byte[24 * 65535];
-    private static readonly int[] _eolBuffer = new int[65535 + 256];
+    private static readonly ushort[] _eolBuffer = new ushort[65535 + 256];
     private static readonly string _chars0LF = File.ReadAllText("Comparisons/Data/65K_Records_Data.csv");
     private static readonly string _chars1LF = File.ReadAllText("Comparisons/Data/SampleCSVFile_556kb_4x.csv");
     private static readonly byte[] _bytes0LF = Encoding.UTF8.GetBytes(_chars0LF);
